Crypto_Trading_Bot_API_Removed/requirements.txt
ccxt>=4.3.89
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
